# library_distributed_server/docs/coordinator/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Coordinator service for distributed transactions in the library management system",
        "title": "Distributed Transaction Coordinator API",
        "termsOfService": "N/A",
        "contact": {
            "name": "API Support",
            "email": "dev@library.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/stats/readers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get readers with borrowing statistics with pagination (enhanced for Flutter app)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get readers with statistics and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number (0-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for reader name or ID",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Readers with stats retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats/system": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get comprehensive system statistics across all sites (Manager only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get system statistics",
                "responses": {
                    "200": {
                        "description": "System statistics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SystemStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied - Manager role required",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.BookWithAvailability": {
            "description": "Book information combined with availability count for client applications",
            "type": "object",
            "properties": {
                "availableCount": {
                    "description": "Available copies count",
                    "type": "integer",
                    "example": 5
                },
                "borrowedCount": {
                    "description": "Currently borrowed count",
                    "type": "integer",
                    "example": 5
                },
                "isbn": {
                    "description": "Book ISBN (ID)",
                    "type": "string",
                    "example": "978-0-123456-78-9"
                },
                "tacGia": {
                    "description": "Author name",
                    "type": "string",
                    "example": "Nguyễn Văn A"
                },
                "tenSach": {
                    "description": "Book title",
                    "type": "string",
                    "example": "Lập trình Go"
                },
                "totalCount": {
                    "description": "Total copies count",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "models.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "details": {
                    "description": "Error details (optional)"
                },
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "models.ListResponse": {
            "description": "Generic paginated list response matching Flutter BookListModel structure",
            "type": "object",
            "properties": {
                "items": {
                    "description": "List of items (matches Flutter items field)"
                },
                "paging": {
                    "description": "Pagination info (matches Flutter paging field)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PagingInfo"
                        }
                    ]
                }
            }
        },
        "models.PagingInfo": {
            "description": "Pagination information matching Flutter PagingModel structure",
            "type": "object",
            "properties": {
                "page": {
                    "description": "Current page number (0-based, matches Flutter)",
                    "type": "integer",
                    "example": 0
                },
                "size": {
                    "description": "Items per page (matches Flutter)",
                    "type": "integer",
                    "example": 20
                },
                "totalPages": {
                    "description": "Total number of pages (matches Flutter)",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "models.SiteStats": {
            "description": "Site-specific statistics",
            "type": "object",
            "properties": {
                "booksOnLoan": {
                    "description": "Books currently on loan",
                    "type": "integer",
                    "example": 75
                },
                "siteID": {
                    "description": "Site identifier",
                    "type": "string",
                    "example": "Q1"
                },
                "totalBooks": {
                    "description": "Total books in this site",
                    "type": "integer",
                    "example": 500
                },
                "totalReaders": {
                    "description": "Total registered readers",
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.SystemStatsResponse": {
            "description": "Comprehensive system statistics across all sites",
            "type": "object",
            "properties": {
                "activeBorrows": {
                    "description": "Currently borrowed books",
                    "type": "integer",
                    "example": 500
                },
                "generatedAt": {
                    "description": "Stats generation time",
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "overdueBooks": {
                    "description": "Overdue books",
                    "type": "integer",
                    "example": 50
                },
                "popularBooks": {
                    "description": "Most borrowed books",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BookWithAvailability"
                    }
                },
                "siteStats": {
                    "description": "Per-site statistics",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SiteStats"
                    }
                },
                "totalBooks": {
                    "description": "Total books in catalog",
                    "type": "integer",
                    "example": 1000
                },
                "totalCopies": {
                    "description": "Total book copies",
                    "type": "integer",
                    "example": 5000
                },
                "totalReaders": {
                    "description": "Total registered readers",
                    "type": "integer",
                    "example": 2000
                }
            }
        },
...
    }
}